Sets up stats tracking
index.tsx now tracks the total number of rolls. The specific % stats are now showing for each RollTracker that is associated with each 2d6 roll. RollTrackerList removes the logic for setting up the 11 RollTracker objects to declutter index.tsx. State for the total number of rolls is passed from index.tsx down to the RollTrackerList and again down to the RollTracker.