Working on modifying the logic for tracking dice rolls to enable clear operation
RollTrackerList is now going to be the source of truth for tracking dice values; this makes it so that the values are easily clearable.

The RollTracker needs to be updated to receive the values from the RollTrackerList, and ultimately should not have any state management outside of informing the RollTrackerList when an operation needs to occur and what kind of operation it is.

Button.tsx slightly increased the hitSlop and added a theme for the clear button -- styling WIP.

index.tsx no longer manages state for the total number of rolls; this is also now in the RollTrackerList.