Still trying to fix
webview is not great; mobile has proper scrolling and displays all elements, but web has 2 cutoff and the expected percentages.

The overhaul for state logic is also not working correctly; the first increment doesn't update the total rolls and the percentages are always 0. I suspect this has something to do how the rendering is actually occurring, and it might be good to find a new way to directly tie the two states to one another -- or perhaps calculate the percentage directly in the RollTracker attribute.