Fixed the styling on the webview
Webview styling resolved by not using alignItems where it wasn't needed, and by adding the flexGrow style to the ScrollView in index.tsx.

Remove several extraneous stylings that were either not being used or were components that didn't need to present.

Fixed the clear button; now properly aligned by adding a view that is centered, and the "Clear" label is now properly contained within the white background.

Percentage and roll counts are now in a fixed position; they will not shift at all as the user increments/decrements the dice values.